% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verstappen.R
\name{did_max_win}
\alias{did_max_win}
\title{Did Verstappen finish first in the most recent f1 event?}
\usage{
did_max_win()
}
\value{
String statement of "Yes" or "No" with country, race event and date.
}
\description{
This function queries the 'openf1' API for Max's last position
in the most recent f1 session (FP1, FP2, FP3, Qualifying, Sprint or Race)
}
